refactor: move screen wrapper and header to parent component
