+++ src/components/AmountTextInput.js
@@ -9,7 +9,7 @@ const propTypes = {
     formattedAmount: PropTypes.string.isRequired,
 
     /** A ref to forward to amount text input */
-    forwardedRef: PropTypes.oneOfType([PropTypes.func, PropTypes.shape({current: PropTypes.instanceOf(React.Component)})]),
+    forwardedRef: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
 
     /** Function to call when amount in text input is changed */
     onChangeAmount: PropTypes.func.isRequired,
