+++ src/components/TextInputWithCurrencySymbol.js
@@ -6,7 +6,7 @@ import * as CurrencyUtils from '../libs/CurrencyUtils';
 
 const propTypes = {
     /** A ref to forward to amount text input */
-    forwardedRef: PropTypes.oneOfType([PropTypes.func, PropTypes.shape({current: PropTypes.instanceOf(React.Component)})]),
+    forwardedRef: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
 
     /** Formatted amount in local currency  */
     formattedAmount: PropTypes.string.isRequired,
